+++ src/main/java/emi/mtg/deckbuilder/view/dialogs/PrintSelectorDialog.java
@@ -58,7 +58,7 @@ public class PrintSelectorDialog extends Dialog<Card.Print> {
 
 		pane.autoEnabled.set(true);
 
-		pane.showingVersionsSeparately.set(true);
+		pane.uniqueness().set(CardView.Uniqueness.Prints);
 		pane.setPrefHeight(scene.getHeight() / 1.5);
 		pane.setPrefWidth(scene.getWidth() / 1.5);